+++ src/Haskellorls/WrappedText.hs
@@ -3,6 +3,7 @@
 module Haskellorls.WrappedText
   ( WrappedText (..),
     wtLength,
+    wtLengthForDisplay,
     render,
     toWrappedText,
     toWrappedTextSingleton,
@@ -11,6 +12,7 @@ module Haskellorls.WrappedText
 where
 
 import qualified Data.Text as T
+import qualified Haskellorls.Utils as Utils
 
 data WrappedText = WrappedText
   { wtPrefix :: T.Text,
@@ -20,6 +22,9 @@ data WrappedText = WrappedText
 
 wtLength :: WrappedText -> Int
 wtLength = T.length . wtWord
+
+wtLengthForDisplay :: WrappedText -> Int
+wtLengthForDisplay = Utils.textLengthForDisplay . wtWord
 
 render :: WrappedText -> T.Text
 render wt = wtPrefix wt <> wtWord wt <> wtSuffix wt